[feature] Enable customizing networking library reducing bundle size
