+++ src/resource/index.ts
@@ -1,4 +1,5 @@
 import Resource from './Resource';
+import SimpleResource from './SimpleResource';
 import {
   SchemaOf,
   DeleteShape,
@@ -43,4 +44,12 @@ export type SchemaList<T> = SchemaList<T>;
 export type SchemaDetail<T> = SchemaDetail<T>;
 export type RequestResource<RS> = RequestResource<RS>;
 
-export { Resource, isEntity, normalize, denormalize, isDeleteShape, schemas };
+export {
+  Resource,
+  SimpleResource,
+  isEntity,
+  normalize,
+  denormalize,
+  isDeleteShape,
+  schemas,
+};
